tests, libnet: Relocate validation to libnet
The validation helper is relocated to the libnet helper package from the
network package. With this change, other tests are no longer depending
on the network tests package.
